DataService/requirements.txt
azure-functions==1.13.3
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
Flask==2.2.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pymongo==4.3.3
requests==2.28.2
urllib3==1.26.15
Werkzeug==2.2.3
